Third Integraton of Alpah Beta variants
This is the first full integration of the alfpha beta variants.
The Scheduler is not working properly, the loop never reaches the task execution, the workaround is to use the 1ms PIT IRQ and execute the 1ms taks there.
Need to analyze if the problem can be solved an use the correct Scheduler.

The movement is dected but with the Computer the voltage suplly makes the EMG to vary to much in a way that is not detectable by the SW. the normal averge voltage is 1300 counts that corresponds to  1v and it doe snto vary to much.
This issue is known and it needs the  board disconnected form the computer, unfurtunatelly is not possible right now there is and issue with the Project set up where the debugging session is taking to much to initialize and once it its disconected the Board wont work in stand alone.  Need to figure out that to do there.

The pedning activities is to create a  Button control to switch between effects.